==> src/Limpieza de datos y Modelización/Código/1_ventas_dep.ipynb <==
### Importación de librerías

import pandas as pd
import numpy as np
import re

### Comienzo de la depuración

df_ventas = pd.read_csv("items_ordered_2years.csv", sep=",")

En este dataframe los duplicados son valores que no tienen valor, puesto que no puede haber dos productos con el mismo item_id. Por tanto, son valores que hay que eliminar

#Eliminación de duplicados
df_ventas = df_ventas.drop_duplicates()

df_ventas.head()

df_ventas.nunique()

Sigue habiendo mas registros que valores únicos en item_id, lo cual indica que hay que revisar qué puede estar generando esa duplicidad de id.

### Revisión de ITEM_ID

df_ventas["item_id"].nunique()

item_nu = df_ventas.groupby(
    "item_id", as_index=False
).size()

mascara = item_nu.iloc[:,1] > 1

item_repetidos = item_nu[mascara]

item_repetidos.head()

mascara2 = df_ventas["item_id"].isin(
    item_repetidos["item_id"].tolist()
)

df_ventas_repetido = df_ventas[mascara2].sort_values("item_id")

df_ventas_repetido.head()

Las duplicidades del item_id aparecen debido a presumiblemente errores en el zipcode/city. Se tratará despues de corregir las columnas de city y zipcode

### Revisión de valores nulos

df_ventas.isna().sum()

Valores nulos en zipcode y city similares. Pueden ser coincidentes.

mascara_na = df_ventas["zipcode"].notna()
df_ventas[mascara_na ].isna().sum()


Efectivamente, al filtrar por zipcodes que no tuvieran na en la columa zipcode desaparecían los valores perdidos en la de city.
Esos registros dificilmente pueden ser imputados por algoritmos como randomforest o knn, sin arriesgarse a estar añadiendo información de forma casi aleatoria.
Por ello, se prescindirá de esos registros.

df_ventas = df_ventas[mascara_na]

### Revisión de zipcodes

df_ventas["longitud_zip"] = df_ventas["zipcode"].dropna().apply(len)
df_ventas["longitud_zip"].unique().tolist()[0:5]

Funciones para limpieza de city y zipcode

#Limpieza de zipcodes con RegEx
def num_guion(string):
    """ Get a string with the numbers and hyphens of another string
    
    Args:
        df: string used to extract the string with numbers abd hyphens

    Returns:
        df: the string with numbers and hyphens
    """
    aux = re.match("([\d-]+)", str(string))
    try:
        return str(aux.group())
    except:
        return string


# Limpieza de nombres de ciudad

# def AcentosLimpiador(text):
# 	acentos = {'ñ':'n','á': 'a', 'é': 'e', 'í': 'i', 'ó': 'o', 'ú': 'u', 'à':'a'}
# 	for ele in acentos:
# 		if ele in text:
# 			text = text.replace(ele, acentos[ele])
# 	return text

# def CityCleaner(text):
#     stopWordSpanish = set(stopwords.words('spanish'))
#     wordTokens = word_tokenize(AcentosLimpiador(text.lower()).rstrip()) 
#     filteredSentence = [element for element in wordTokens if not element in stopWordSpanish] 
#     return filteredSentence

    


Limpieza de ambas columnas para facilitar el merge con la tabla de zipcodes

# Limpieza zipcode
df_ventas["zipcode"] = df_ventas["zipcode"].apply(lambda x: num_guion(x))
# df_ventas["city"] = df_ventas["city"].apply(lambda x: CityCleaner(x))
# df_ventas["city"] = df_ventas["city"].apply(lambda x: "".join(x))

Se carga el dataframe con la información de privincias, países, zipcodes y ciudades.

zipcodes_csv = pd.read_csv("DataScrapped.csv")
zipcodes_csv.drop("Unnamed: 0", axis=1, inplace=True)
zipcodes_csv.drop_duplicates(inplace=True)


zipcodes_csv.head()

Se genera una copia del dataframe con las columnas city y zipcode para un primer join/merge

# Nombres de las columnas a minuscula

column_list = zipcodes_csv.columns.tolist()
for i in range(len(column_list)):
    column_list[i] = column_list[i].lower()
zipcodes_csv.columns = column_list

# Obtener dataframe con solo city y zipcode, y pasar city a minusculas para facilitar el match

# zipcodes_to_join = zipcodes_csv.loc[:, ["zipcode", "city"]]
# zipcodes_to_join["city"] = zipcodes_to_join["city"].apply(lambda x: CityCleaner(x))
# zipcodes_to_join["city"] = zipcodes_to_join["city"].apply(lambda x: "".join(x))

Para obtener aquellos zipcodes que estén mal escritos pero que estén bien la ciudad, se realiza un primer merge por columna city


df_ventas.drop("city", inplace=True, axis=1)

df_ventas_2 = df_ventas.merge(zipcodes_csv, how="left", on="zipcode")

df_ventas_2.head()

mascara_null_2 = df_ventas_2["city"].notna()
df_ventas_2 = df_ventas_2[mascara_null_2]


df_ventas_2.drop_duplicates(inplace=True)

### Solucionar el problema de id duplicadas

Se parte de la hipótesis de que son registros que se han duplicado al hacerse una corrección sobre la ubicación donde se mandaba el producto pero que no se eliminó el registro anerior. No se puede saber cual es el correcto, por lo que se decide tomar uno de los dos valores de forma alaetoria, que solo afectará a la ciudad/zipcode, mateniendose la comunidad y país de igual manera.

#Diccionario de la función de agregación por campo

list_columns = df_ventas_2.drop("item_id", axis = 1).columns.to_list()
list_agg = ["first"] * len(list_columns)
dict_agg = dict(zip(list_columns, list_agg))

# Nos deshacemos de los duplicados

df_ventas_3 = df_ventas_2.groupby(
    "item_id", as_index=False
).agg(dict_agg)

df_ventas_3.head()

#Exportación de ventas_1

df_ventas_3.to_csv('df_ventas_1.csv')
