==> src/Limpieza de datos y Modelización/Código/2_dates_and_amounts.ipynb <==
### Importación de librerías

import pandas as pd
import numpy as np
import datetime as dt


# Carga del dataframe de ventas depurado

df_ventas = pd.read_csv("df_ventas_1.csv", sep=",")

### Tratamiento del campo hora para obtener variables de cara a la visualización/modelos

# Obtención de distintos campos de fecha

# Campo sin horas
df_ventas["date"] = df_ventas["created_at"].apply(lambda x: x[0:10])

# Capo de año
df_ventas["year"] = df_ventas["created_at"].apply(lambda x: x[0:4])

# Campo de horas por franja
df_ventas["hour"] = df_ventas["created_at"].apply(lambda x: x[11:13])


# Campo de numero de semana (del 1 al 104)
df_ventas["week"] = df_ventas["date"]
df_ventas["week"] = df_ventas["week"].apply(lambda _: dt.datetime.strptime(_,"%Y-%m-%d"))
df_ventas["week"] = df_ventas["week"].dt.isocalendar().week
df_ventas["week"][df_ventas["year"] == "2018"] = df_ventas["week"][df_ventas["year"] == "2018"] + 52

# Campo dia de la semana
df_ventas["date"] = pd.to_datetime(df_ventas["date"])
df_ventas["day"] = df_ventas["date"].apply(lambda x: pd.Timestamp(x).day_name())

df_ventas["margin_total"] = (df_ventas["price"] - df_ventas["base_cost"]) * df_ventas["qty_ordered"]
df_ventas["price_total"] = df_ventas["price"] * df_ventas["qty_ordered"]

df_ventas.head()

df_ventas.drop("Unnamed: 0", inplace=True, axis = 1)

**Exportación de resultados a formato CSV**

df_ventas.to_csv('df_ventas_2.csv')
