==> src/Limpieza de datos y Modelización/Código/3_products_categories.ipynb <==
### Importación de librerías

import pandas as pd
import numpy as np
import re

### Inicio de la depuración

products = pd.read_csv("products_limpieza1.csv")
df_categories = pd.read_csv("categories_limpieza1.csv")

products.head()

df_categories.head(6)

**Fusión de dataframes**

products_full = pd.merge(products, df_categories, on='sku', how='left')

products_full.head(6)

products_full["cat1"].value_counts()

products_full["analytic_category"].value_counts()

#Para unificar categorias pasamos a lower (minúsculas):
products_full["analytic_category"] = products_full["analytic_category"].str.lower()

products_full["analytic_category"].head()

products_full[products_full["analytic_category"].isna()].head()

products_full.info()

# Por qué esas variables unnamed:0_? Se eliminan...

products_full.drop(columns=["Unnamed: 0_x", "Unnamed: 0_y"]).head()

### Es necesario que "analytic_category" tome lo que está en el mismo registro en "cat1"

products_full["analytic_category"].isna().sum()

(products_full["analytic_category"] == None).sum()

for pos, na in enumerate(products_full["analytic_category"]): #no funciona... ¡?
    if na == None:
        products_full["analytic_category"].iloc[pos] = products_full["cat1"].iloc[pos]

products_full.info()

products_full["analytic_category"].isna().sum()

**Revisión de tipos**

products.info()

df_categories.info()

#test2 = [re.search("([^;]*)", element).group() for element in categories_sep["cat2,3"]]

**Eliminación de columnas no relevantes**

products_full.drop(columns=["Unnamed: 0_x", "Unnamed: 0_y"], inplace=True, axis=1)

**Exportación de resultados a CSV**

products_full.to_csv("df_products_dep.csv")
