==> src/Limpieza de datos y Modelización/Código/4_ventas_def.ipynb <==
### Importación de librerías

import pandas as pd
import numpy as np
import sys

## Inicio de la depuración

# Carga de df
df_ventas = pd.read_csv("df_ventas_2.csv")
df_zipcodes = pd.read_csv("df_products_dep.csv")

# Se elimina la columna derivada de los indices
df_ventas.drop(columns="Unnamed: 0",inplace=True)
df_zipcodes.drop(columns="Unnamed: 0",inplace=True)

df_ventas_full = df_ventas.merge(df_zipcodes, how="left", on="product_id")

Se hace drop de las categorías, ya que la dos y tres contienen muchos valores perdidos, y la cat1 se encuentra recogida en analytic_category

# 
df_ventas_full.drop(columns = df_ventas_full.loc[:, ["cat1", "cat2", "cat3"]], inplace=True, axis=1)

df_ventas_full.isna().sum()

df_ventas_full["product_id"] = df_ventas_full["product_id"].astype("float", errors="ignore")

Borramos columnas con poco valor para visualización y modelización que además contienen valores perdidos y que su imputación estaría expuesta a ser de forma semi aleatoria.

df_ventas_full.drop(columns = df_ventas_full.loc[:, ["sku", "short_description", "picture"]], inplace=True, axis=1)

**Exportación final de resultados**

df_ventas_full.to_csv("df_ventas_3.csv")

Imputación por randomforest de los valores perdidos mediante la librería de missingpy
